improved smart contract code base
