+++ afya_backend/src/backend/utils/validateUser.ts
@@ -1,0 +1,6 @@
+import { text } from "azle/experimental";
+import { users } from "../storage/storage";
+
+export function validateUser(userId: text): boolean {
+    return users.get(userId) !== undefined;
+}
